Week3/Day2_AI_Agent_Basics/Task1_MultiAgent_RAG_System/requirements (2).txt
streamlit
sentence-transformers
faiss-cpu
langchain
langchain-community
google-genai


